♻️ Move UncommittedChanges and UntrackedFiles to git package
